pyproject.toml
[tool.poetry]
name = "pytorres"
version = "0.1.0"
description = ""
authors = ["zwanto <antoine.hamon@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.87.0"
uvicorn = {extras = ["standard"], version = "^0.19.0"}
elasticsearch = "^8.5.0"
python-dotenv = "^0.21.0"
parse-torrent-title = "^2.4"
requests = "^2.28.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.1.9"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
